refactor file upload to service and persistence structure for re-use
